Pretty big push all in one, here we go:
Started an entirely new look for the site, based on suggestions from commissioner. Still same basic functionality but a softer look, friendlier home pagel, friendlier shop experience, drop down cart modals, better organization for products, drop down for category list. Add animations, animations make sites look better;

Initially attempted to change the navigation because of the introduction of more serious header and footer, but realized it would take too much effort at this point. I wanted a router element in app.html that was smart about mobile or desktop, and then within either of those, have another router that routed pages sandwiched between a permanent header and footer. Too complicated, realized I could do the same with just holding a view variable in service and ngIfing the different components between header and footer. The result is a site that only ever loads one page. Works for mobile but seems somewhat unnatural on browser. To be determined if that’s the best practice;

So, added a footer component, an info page component, and a homepage component;

To make a better category and subcategory filter, I simplified the hierarchy and made the app work with just categories and subcategories rather than aisles;

Updated the api endpoint to match the new dns of the Django server. The frontend app and the Django server are currently running on the same server. This probably needs to change;

Updated index.html to include the correct web fonts, instantly making the site look 100% better;

I completely re-vamped the shop page component and the header component.
Gave the header a search bar, as well as an animated cart list modal on hover of cart button, and an animated drop down select of categories of items on hover over shop link.
During this process I discovered the following:

Some trickery with css up at the top navbar.
What’s the best way to make a good hover path between a menu hover link and the drop down that animates below it?
You can’t really use ngif or hidden if you want css animations because apparently they need to be in the Dom for the animations to work
I did come upon a cool little method of dynamically activating an animation based on some other action happening elsewhere pn the page, not necessarily the div where the animation should happen. This allows me to animate a menu to fade in from bottom of screen based on hovering over a navigation link.
But so I ended up trying to have this parent, relative position element that has the link in it, and then have an ‘animation’ container div that’s absolutely positioned within that parent relative element and then within that is a statically positioned div with the content in it, on hover over the link, the innermost div which starts with a margin up top to make it low in the screen then transitions up with a translation
Ran into issues then having mouse leave and enter events on the animated menu, fixed it by disabling pointer events on it until it was ‘activated’ by hovering over link
Also an issue with the link and animation container in same div cause the hover events of the link to be eclipsed by the other div. this I ficxesd by making the link also absolutely positioned within the relative parent and giving it a higher z index

Is there a better way?
Should it be this hard?

In terms of the shop page:

I wanted a side nav that popped in and out, and within it, a tree structure to represent selection of categories and subcategories. Initially attempted to do this with angular material side navbars and material trees, but these things add so much bloat for no reason, it was easier to just make it from scratch.
In terms of the product view, added the option to view a grid or list, so had to build both those views. I discovered the magic of css grid styling. Just use a grid. So simple. For list items, grids give each element a certain amount of space, and the list item expands to fit the page. With the grid, the elements themselves can be given dimensions, and then set the grid to a single, auto-fitting column, this allows the grid to responsively add columns according to the size of screen. I found with images, it was necessary to give them specific dimensions to preserve aspect ratios and make the image look good, and then have the element containing it be a certain dimension based on that. So the image was the source of dimension.

Also added an entirely new item detail view when you click on the item in the list.

Still need to:

Make the cart modal work better
Move the item detail view to be within the shop grid area to make user more comfortable
Animate the category slide
The category select on the shop is buggy, the subcategories won’t select
Do the info page.
Do the home page banners etc, which includes animating the slide at the top.