Product list items looked too squished, spaced them out. Fixed spacing for product grid items. Slowed animation speed down for hover product items. Way more satisfying. Speed up animations for cart modal and shop list. Fixed colors on cat expand butt and '1 in cart' div.
Refactored the server and app code to have a more simplified system for product price and measurement. Measurement for now will now be listed in the name, like Amazon. Cleaned up the previous hacky way of changing the category system as well. Now just categories and subcategories.

Still work to do on changing button styles, and making links work correctly between the header and the shop page.