Started implementing searchbar. Fixed some css with it.
Was not able to get the funtionality to display a pending/filler/pagination div based on if the search was currently "searching", like Reddit. Wanted to make it like Reddit with if the user searches something, it displays an intermediate animation while it's 'loading' data from the server. This is just an aesthetic decision currently though, seeing as the app has no data to load from the server, and with Reddit, it's probably searching billions of records, and with this site, it's searching less than 50. The main reason couldn't implement was some angular issues with the dom and having ngifs with 'searching' variables. may instead in the future if want to implement use subscriptions and promises, since some of the lodash filtering is async.
Reddit also displays things like 'trending today' if the user hasn't searched anythign yet.

Noticed an error with ngModelChange getting called before the ngModel is changed, you can fix this by making the ngModelChange appear after ngModel in the html line :|

Also, the filter is currently stringifying all json objects and searching for characters which means it's getting items that having matchign characters in the field keys which shouldn't happen. can fix later.

There is a problem with the search dropdown scrolling being too easily overwritten by global page scrolling. However, checked if this happened on Reddit and it does.