Made info bottoms padding less
Fixed styling issue with cat side panel and also logic for showing selected subcat.

Fixed styling for shop add to cart buttons

Started work on the storefront page and animating the slider:

Was able to get the first text to slide on and off the screen using margins and transitions except there was this weird issue with the text getting squashed as it entered and left the screen and “entered” or “exited” its parent element. Temporary fix by giving a fixed height and width to the text block and setting overflow to auto, but might want to make an absolute div in the future, because it also now has a brief moment where it looks like it’s getting covered by an invisible div. want it to be animating on top of the whole page.

Todo:
Make the shop list item different. Only clickable if not on add to cart or quantity change.
Need to fix category side panel width to be smaller
Need to fix remove from cart button on product view and styling of text