It’s not really scaleable. It works basically now for two slides but if you were to add slides it would complicate things
I have a method for doing an animation back and forth on an element based on a variable, but I have no way to, for example, when come into being an element transform it from left, but then when it’s deleted, instead of doing the reverse of transition back to left, just do an opacity fade. So it’s only one way back and forth.

Might be better in the future to make each slide an angular component that animates or deanimates on create and destroy, but I’ve run into issues with ngIfs and [hidddens] with css animations in the past

The background image transition is not ideal. There is obviously a much better way to be doing a slider with css, but for now this works.

Bug with button-10 not having hover, override with !important, not ideal.

Todo:
Add a timed loop that animates the banner if the user hasn’t interacted with the site in x seconds.