Some design notes:
No easy way to navigate back to the page you were just on. Seems like pretty intuitive for a user to be able to do that. I always want to quickly be able to go back. I guess that’s what routing is for? Maybe that’s the next step, to make it do routing properly, instead of a single-page app kind of thing. The users can navigate back in the browser.

On the way, learned some stuff:

Some strange css transition delays encountered. General solution news I was animating all styles in the css class, which adds overhead. Only animate the specific styles you need, like margin bottom or something

Implemented various side or bottom sliding panels, used absolute positioned masks to allow for clicking out of the container, and ran into troubles with horizontal scrolling.

In the process of using masks and flex containers with things margined offscreen initially that slide into view, I discovered a more intuitive way to make a banner slider, and so implemented that on the mobile page. The animations of the texts overlayed on each banner slide were too finicky and fine-motor-skill-ish, so left them out for now.

Todo:
Add back in the animations for the mobile banner slide texts.
Still need to fix some basic padding, and make the horizontal scroll for the day of delivery select better.
A little phantom cart modal pops up when you click the cart, need to get rid of it
Need to add extra delivery info form in checkout.
Styling on the shop list needs fixing